CVSL-2160: Move address check request endpoints to AddressResource, fix return type in schema doc.
