+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsassessforearlyreleaseapi/resource/OffenderResource.kt
@@ -15,27 +15,21 @@ import org.springframework.http.MediaType
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.PutMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
 import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.config.ErrorResponse
-import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.model.AddCasCheckRequest
-import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.model.AddStandardAddressCheckRequest
 import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.model.AssessmentSummary
 import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.model.OffenderSummary
 import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.model.OptOutReasonType
 import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.model.OptOutRequest
-import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.model.StandardAddressCheckRequestSummary
 import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.service.OffenderService
-import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.service.prison.AddressService
 
 @RestController
 @RequestMapping(produces = [MediaType.APPLICATION_JSON_VALUE])
 class OffenderResource(
-  private val addressService: AddressService,
   private val offenderService: OffenderService,
 ) {
 
@@ -179,98 +173,4 @@ class OffenderResource(
 
     offenderService.optOut(prisonNumber)
   }
-
-  @PostMapping("/offender/{prisonNumber}/current-assessment/standard-address-check-request")
-  @PreAuthorize("hasAnyRole('ASSESS_FOR_EARLY_RELEASE_ADMIN')")
-  @ResponseStatus(code = HttpStatus.CREATED)
-  @Operation(
-    summary = "Adds a standard address check request for an offender.",
-    description = "Adds a standard address check request for an offender's current assessment.",
-    security = [SecurityRequirement(name = "assess-for-early-release-admin-role")],
-  )
-  @ApiResponses(
-    value = [
-      ApiResponse(
-        responseCode = "201",
-        description = "The standard address check request has been added to the offender's current assessment.",
-        content = [
-          Content(
-            mediaType = "application/json",
-            schema = Schema(implementation = StandardAddressCheckRequestSummary::class),
-          ),
-        ],
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorised, requires a valid Oauth2 token",
-        content = [
-          Content(
-            mediaType = "application/json",
-            schema = Schema(implementation = ErrorResponse::class),
-          ),
-        ],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Forbidden, requires an appropriate role",
-        content = [
-          Content(
-            mediaType = "application/json",
-            schema = Schema(implementation = ErrorResponse::class),
-          ),
-        ],
-      ),
-    ],
-  )
-  fun addStandardAddressCheckRequest(
-    @Parameter(required = true) @PathVariable prisonNumber: String,
-    @Valid @RequestBody addStandardAddressCheckRequest: AddStandardAddressCheckRequest,
-  ) = addressService.addStandardAddressCheckRequest(prisonNumber, addStandardAddressCheckRequest)
-
-  @PostMapping("/offender/{prisonNumber}/current-assessment/cas-check-request")
-  @PreAuthorize("hasAnyRole('ASSESS_FOR_EARLY_RELEASE_ADMIN')")
-  @ResponseStatus(code = HttpStatus.CREATED)
-  @Operation(
-    summary = "Adds a CAS check request for an offender.",
-    description = "Adds a CAS check request for an offender's current assessment.",
-    security = [SecurityRequirement(name = "assess-for-early-release-admin-role")],
-  )
-  @ApiResponses(
-    value = [
-      ApiResponse(
-        responseCode = "201",
-        description = "The CAS check request has been added to the current assessment for the offender.",
-        content = [
-          Content(
-            mediaType = "application/json",
-            schema = Schema(implementation = StandardAddressCheckRequestSummary::class),
-          ),
-        ],
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorised, requires a valid Oauth2 token",
-        content = [
-          Content(
-            mediaType = "application/json",
-            schema = Schema(implementation = ErrorResponse::class),
-          ),
-        ],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Forbidden, requires an appropriate role",
-        content = [
-          Content(
-            mediaType = "application/json",
-            schema = Schema(implementation = ErrorResponse::class),
-          ),
-        ],
-      ),
-    ],
-  )
-  fun addCasCheckRequest(
-    @Parameter(required = true) @PathVariable prisonNumber: String,
-    @Valid @RequestBody addCasCheckRequest: AddCasCheckRequest,
-  ) = addressService.addCasCheckRequest(prisonNumber, addCasCheckRequest)
 }
