+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsassessforearlyreleaseapi/integration/OffenderResourceIntTest.kt
@@ -9,18 +9,13 @@ import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.core.ParameterizedTypeReference
 import org.springframework.test.context.jdbc.Sql
-import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.entity.AddressPreferencePriority
 import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.entity.AssessmentStatus
 import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.integration.base.SqsIntegrationTestBase
 import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.integration.wiremock.PrisonRegisterMockServer
-import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.model.AddCasCheckRequest
-import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.model.AddStandardAddressCheckRequest
 import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.model.AssessmentSummary
-import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.model.CasCheckRequestSummary
 import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.model.OffenderSummary
 import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.model.OptOutReasonType
 import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.model.OptOutRequest
-import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.model.StandardAddressCheckRequestSummary
 import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.repository.AssessmentRepository
 import uk.gov.justice.digital.hmpps.hmppsassessforearlyreleaseapi.repository.OffenderRepository
 import java.time.LocalDate
@@ -30,8 +25,6 @@ private const val PRISON_NUMBER = "A1234AD"
 private const val GET_CASE_ADMIN_CASELOAD_URL = "/prison/$PRISON_CODE/case-admin/caseload"
 private const val GET_CURRENT_ASSESSMENT_URL = "/offender/$PRISON_NUMBER/current-assessment"
 private const val OPT_OUT_ASSESSMENT_URL = "/offender/$PRISON_NUMBER/current-assessment/opt-out"
-private const val ADD_STANDARD_ADDRESS_CHECK_REQUEST_URL = "/offender/$PRISON_NUMBER/current-assessment/standard-address-check-request"
-private const val ADD_CAS_CHECK_REQUEST_URL = "/offender/$PRISON_NUMBER/current-assessment/cas-check-request"
 
 class OffenderResourceIntTest : SqsIntegrationTestBase() {
 
@@ -227,135 +220,6 @@ class OffenderResourceIntTest : SqsIntegrationTestBase() {
         .expectStatus()
         .isBadRequest
     }
-  }
-
-  @Nested
-  inner class AddStandardAddressCheckRequestTests {
-    private val caInfo = "ca info"
-    private val ppInfo = "pp info"
-    private val priority = AddressPreferencePriority.THIRD
-    private val uprn = "200010019924"
-
-    @Test
-    fun `should return unauthorized if no token`() {
-      webTestClient.post()
-        .uri(ADD_STANDARD_ADDRESS_CHECK_REQUEST_URL)
-        .bodyValue(anAddStandardAddressCheckRequest())
-        .exchange()
-        .expectStatus()
-        .isUnauthorized
-    }
-
-    @Test
-    fun `should return forbidden if no role`() {
-      webTestClient.post()
-        .uri(ADD_STANDARD_ADDRESS_CHECK_REQUEST_URL)
-        .headers(setAuthorisation())
-        .bodyValue(anAddStandardAddressCheckRequest())
-        .exchange()
-        .expectStatus()
-        .isForbidden
-    }
-
-    @Test
-    fun `should return forbidden if wrong role`() {
-      webTestClient.post()
-        .uri(ADD_STANDARD_ADDRESS_CHECK_REQUEST_URL)
-        .headers(setAuthorisation(roles = listOf("ROLE_WRONG")))
-        .bodyValue(anAddStandardAddressCheckRequest())
-        .exchange()
-        .expectStatus()
-        .isForbidden
-    }
-
-    @Sql(
-      "classpath:test_data/reset.sql",
-      "classpath:test_data/some-offenders.sql",
-      "classpath:test_data/an-address.sql",
-    )
-    @Test
-    fun `should add a standard address check request`() {
-      val addressCheckRequest = webTestClient.post()
-        .uri(ADD_STANDARD_ADDRESS_CHECK_REQUEST_URL)
-        .headers(setAuthorisation(roles = listOf("ASSESS_FOR_EARLY_RELEASE_ADMIN")))
-        .bodyValue(anAddStandardAddressCheckRequest())
-        .exchange()
-        .expectStatus()
-        .isCreated
-        .expectBody(object : ParameterizedTypeReference<StandardAddressCheckRequestSummary>() {})
-        .returnResult().responseBody!!
-
-      assertThat(addressCheckRequest.caAdditionalInfo).isEqualTo(caInfo)
-      assertThat(addressCheckRequest.ppAdditionalInfo).isEqualTo(ppInfo)
-      assertThat(addressCheckRequest.address.uprn).isEqualTo(uprn)
-    }
-
-    private fun anAddStandardAddressCheckRequest(): AddStandardAddressCheckRequest =
-      AddStandardAddressCheckRequest(caInfo, ppInfo, priority, uprn)
-  }
-
-  @Nested
-  inner class AddCasCheckRequestTests {
-    private val caInfo = "ca info"
-    private val ppInfo = "pp info"
-    private val priority = AddressPreferencePriority.THIRD
-
-    @Test
-    fun `should return unauthorized if no token`() {
-      webTestClient.post()
-        .uri(ADD_CAS_CHECK_REQUEST_URL)
-        .bodyValue(aAddCasCheckRequest())
-        .exchange()
-        .expectStatus()
-        .isUnauthorized
-    }
-
-    @Test
-    fun `should return forbidden if no role`() {
-      webTestClient.post()
-        .uri(ADD_CAS_CHECK_REQUEST_URL)
-        .headers(setAuthorisation())
-        .bodyValue(aAddCasCheckRequest())
-        .exchange()
-        .expectStatus()
-        .isForbidden
-    }
-
-    @Test
-    fun `should return forbidden if wrong role`() {
-      webTestClient.post()
-        .uri(ADD_CAS_CHECK_REQUEST_URL)
-        .headers(setAuthorisation(roles = listOf("ROLE_WRONG")))
-        .bodyValue(aAddCasCheckRequest())
-        .exchange()
-        .expectStatus()
-        .isForbidden
-    }
-
-    @Sql(
-      "classpath:test_data/reset.sql",
-      "classpath:test_data/some-offenders.sql",
-      "classpath:test_data/an-address.sql",
-    )
-    @Test
-    fun `should add a standard address check request`() {
-      val addressCheckRequest = webTestClient.post()
-        .uri(ADD_CAS_CHECK_REQUEST_URL)
-        .headers(setAuthorisation(roles = listOf("ASSESS_FOR_EARLY_RELEASE_ADMIN")))
-        .bodyValue(aAddCasCheckRequest())
-        .exchange()
-        .expectStatus()
-        .isCreated
-        .expectBody(object : ParameterizedTypeReference<CasCheckRequestSummary>() {})
-        .returnResult().responseBody!!
-
-      assertThat(addressCheckRequest.caAdditionalInfo).isEqualTo(caInfo)
-      assertThat(addressCheckRequest.ppAdditionalInfo).isEqualTo(ppInfo)
-      assertThat(addressCheckRequest.preferencePriority).isEqualTo(priority)
-    }
-
-    private fun aAddCasCheckRequest(): AddCasCheckRequest =
-      AddCasCheckRequest(caInfo, ppInfo, priority)
   }
 
   private companion object {
